# core-banking/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Core Banking Application.",
        "title": "Core Banking API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/accounts/{id}/transactions": {
            "get": {
                "description": "Iterates paginated lists isolated for an account.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List Transactions (By Account)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limits response objects",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination Base64 Cursor",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination Direction (next, prev)",
                        "name": "direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transaction.ResponseSuccessTransactionList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transaction.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "description": "Aggregates the global banking journal history explicitly sorted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List All Transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limits response objects",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination Base64 Cursor",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transaction.ResponseSuccessTransactionList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transaction.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1.0/transfer-intrabank": {
            "post": {
                "description": "Safely processes an inter-account fund transfer linearly.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Account Transfer (Optimistic)",
                "parameters": [
                    {
                        "type": "string",
                        "default": "PARTNER001",
                        "description": "Partner ID provided by the bank",
                        "name": "X-PARTNER-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "2026-04-12T18:00:00Z",
                        "description": "ISO-8601 Timestamp",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "valid-signature-for-testing",
                        "description": "HMAC-SHA256 Signature",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1234567890",
                        "description": "Random unique ID representing the request",
                        "name": "X-EXTERNAL-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Transfer Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transaction.IntrabankTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transaction.IntrabankTransferResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transaction.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transaction.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1.0/transfer-intrabank-locked": {
            "post": {
                "description": "Safely processes a transfer explicitly restricting concurrency on the accounts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Account Transfer (Pessimistic Redis Lock)",
                "parameters": [
                    {
                        "type": "string",
                        "default": "PARTNER001",
                        "description": "Partner ID provided by the bank",
                        "name": "X-PARTNER-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "2026-04-12T18:00:00Z",
                        "description": "ISO-8601 Timestamp",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "valid-signature-for-testing",
                        "description": "HMAC-SHA256 Signature",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1234567890",
                        "description": "Random unique ID representing the request",
                        "name": "X-EXTERNAL-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Transfer Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transaction.IntrabankTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transaction.IntrabankTransferResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transaction.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transaction.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1.0/transfer/status": {
            "post": {
                "description": "Fetch transaction status conforming to SNAP BI using Service Code 36.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Transfer Status Inquiry",
                "parameters": [
                    {
                        "type": "string",
                        "default": "PARTNER001",
                        "description": "Partner ID provided by the bank",
                        "name": "X-PARTNER-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "2026-04-12T18:00:00Z",
                        "description": "ISO-8601 Timestamp",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "valid-signature-for-testing",
                        "description": "HMAC-SHA256 Signature",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1234567890",
                        "description": "Random unique ID representing the request",
                        "name": "X-EXTERNAL-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Inquiry Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transaction.TransferStatusInquiryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transaction.TransferStatusInquiryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transaction.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/transaction.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "snap.SNAPAmount": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "value": {
                    "description": "e.g., \"10000.00\"",
                    "type": "string"
                }
            }
        },
        "transaction.ErrorResponse": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
        "transaction.IntrabankTransferRequest": {
            "type": "object",
            "properties": {
                "additionalInfo": {},
                "amount": {
                    "$ref": "#/definitions/snap.SNAPAmount"
                },
                "beneficiaryAccountNo": {
                    "type": "string"
                },
                "customerReference": {
                    "type": "string"
                },
                "feeType": {
                    "type": "string"
                },
                "originatorInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transaction.OriginatorInfo"
                    }
                },
                "partnerReferenceNo": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "sourceAccountNo": {
                    "type": "string"
                },
                "transactionDate": {
                    "type": "string"
                }
            }
        },
        "transaction.IntrabankTransferResponse": {
            "type": "object",
            "properties": {
                "additionalInfo": {},
                "amount": {
                    "$ref": "#/definitions/snap.SNAPAmount"
                },
                "beneficiaryAccountNo": {
                    "type": "string"
                },
                "beneficiaryBankCode": {
                    "type": "string"
                },
                "customerReference": {
                    "type": "string"
                },
                "originatorInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transaction.OriginatorInfo"
                    }
                },
                "partnerReferenceNo": {
                    "type": "string"
                },
                "referenceNo": {
                    "type": "string"
                },
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "sourceAccountNo": {
                    "type": "string"
                },
                "traceNo": {
                    "type": "string"
                },
                "transactionDate": {
                    "type": "string"
                }
            }
        },
        "transaction.OriginatorInfo": {
            "type": "object",
            "properties": {
                "originatorBankCode": {
                    "type": "string"
                },
                "originatorCustomerName": {
                    "type": "string"
                },
                "originatorCustomerNo": {
                    "type": "string"
                }
            }
        },
        "transaction.ResponseSuccessTransactionList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transaction.TransactionHistoryResponse"
                    }
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": true
                },
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
        "transaction.TransactionHistoryResponse": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "accountNumber": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "balance_after": {
                    "type": "integer"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "entry_type": {
                    "type": "string"
                },
                "journal_type": {
                    "type": "string"
                },
                "ledger_entry_id": {
                    "type": "string"
                },
                "partnerReferenceNo": {
                    "type": "string"
                },
                "referenceNo": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "transaction_type": {
                    "type": "string"
                }
            }
        },
        "transaction.TransferStatusInquiryRequest": {
            "type": "object",
            "properties": {
                "originalPartnerReferenceNo": {
                    "type": "string"
                },
                "originalReferenceNo": {
                    "type": "string"
                },
                "serviceCode": {
                    "type": "string"
                }
            }
        },
        "transaction.TransferStatusInquiryResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "$ref": "#/definitions/snap.SNAPAmount"
                },
                "latestTransactionStatus": {
                    "type": "string"
                },
                "originalPartnerReferenceNo": {
                    "type": "string"
                },
                "originalReferenceNo": {
                    "type": "string"
                },
                "referenceNumber": {
                    "type": "string"
                },
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "serviceCode": {
                    "type": "string"
                },
                "transactionStatusDesc": {
                    "type": "string"
                }
            }
        }
    }
}